Update Rapp_means_Vari_update
*UPDATED VERSION*

A for loop was implemented, in this way the program is able 
to fill more than one excel file at a time.
When a new user is inserted into the .csv file, the following 
lists need to be updated:
	- The list of user names
	- The list of excel files where the information'll be stored

the number of times the programme is executed, is determined by 
the number of users in the.csv file. This (numerical) value must be 
updated before the for loop.

In later versions, for loop should be updated and 
if possible, improved.