Update Rapp_means_Vari_update
the program now automatically creates an excel file from an existing template file. Create excel files by hand is no longer necessary. 
In case of problems due to the modification of the program, it is no longer necessary to eliminate every time the pre-existing excel file.
It remains, however, necessary to modify the excel name list when a new csv. file is read