Update Rapp_means_Vari_update
*UPDATED VERSION*
The program now gains much more memory space and speed creating a dataframe with added limitations:
	- ['Valore'] == 0
	- ['Indice'] != 0
in this way df analyzed are smaller and are treated faster. Almost complete dataframes gain little space ( Montalbano SA - 228 had a memory usage of 14.9 MB, now the memory usage is 12.8 MB).
If an analysed table is empty, the reduced df is discarded and is recalculated with the classic method without the added limitations.
In theory the program should be able to calculate fully full annual tables.