Update Rapp_means_Vari_update
*UPDATED VERSION*
The program is now able to automatically create the output excel file. The excel list is no longer needed. It is necessary however indicate the output file path in the dedicated variable in the user section.

A counter was added in order to determine the time taken for each user examined in the program.