Create Automatic mail sender
This program is the extension of the other program. Once the first process has finished to create all the excel 
files, the secondo proces will start. The program searches within a list of users for names that match 
those in excel files. Once the matches are found, the files will be sent to their respective owners.

You need to keep your user list constantly updated, with the right names and the right emails.