Update Rapp_means_Vari_update
*UPDATED VERSION*

Fixed a problem in which if a month without values was selected, in the monthly summary table were only displayed zeros instead of calculated values. Now that table is calculated by omitting the pre-set limitations from the filling of the other tables. this means that even if the month is completely empty (of value), the other months may be filled, if they're not empty (when, for exemple, the values of column "indice" are different from 0).